Fix #3 Mapstruct, DTOs

+++ src/main/kotlin/com/github/balcon/kotlincruddemo/repository/AuthorRepository.kt
@@ -4,9 +4,9 @@ import com.github.balcon.kotlincruddemo.model.Author
 import org.springframework.data.jpa.repository.EntityGraph
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface AuthorRepository : JpaRepository<Author, Int> {
+interface AuthorRepository : JpaRepository<Author, Long> {
     @EntityGraph(attributePaths = ["books"])
-    fun findAuthorWithBooksById(id: Int): Author?
+    fun findAuthorWithBooksById(id: Long): Author?
 
-    fun findAuthorById(id: Int): Author?
+    fun findAuthorById(id: Long): Author?
 }
