+++ src/main/kotlin/com/github/balcon/kotlincruddemo/repository/BookRepository.kt
@@ -4,13 +4,13 @@ import com.github.balcon.kotlincruddemo.model.Book
 import org.springframework.data.jpa.repository.EntityGraph
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface BookRepository : JpaRepository<Book, Int> {
+interface BookRepository : JpaRepository<Book, Long> {
     @EntityGraph(attributePaths = ["author"])
-    fun findBookWithAuthorById(id: Int): Book?
+    fun findBookWithAuthorById(id: Long): Book?
 
-    fun findBookById(id: Int): Book?
+    fun findBookById(id: Long): Book?
 
-    fun findBookByAuthorId(authorId: Int): MutableList<Book>
+    fun findBookByAuthorId(authorId: Long): MutableList<Book>
 
     @EntityGraph(attributePaths = ["author"])
     override fun findAll(): MutableList<Book>
